refactor(glossary): rename temporary `glossary_new` -> `glossary`
ref: #374
