+++ lib/src/authentication/landing_view_model.dart
@@ -4,7 +4,7 @@ import "package:go_router/go_router.dart";
 import "package:kana_to_kanji/src/core/constants/authentication_method.dart";
 import "package:kana_to_kanji/src/core/repositories/user_repository.dart";
 import "package:kana_to_kanji/src/core/services/resources/sync_service.dart";
-import "package:kana_to_kanji/src/glossary_new/view.dart";
+import "package:kana_to_kanji/src/glossary/view.dart";
 import "package:kana_to_kanji/src/locator.dart";
 import "package:rive/rive.dart";
 import "package:stacked/stacked.dart";
