+++ lib/src/glossary/widgets/search_bar.dart
