+++ lib/src/splash/splash_view_model.dart
@@ -1,7 +1,7 @@
 import "package:go_router/go_router.dart";
 import "package:kana_to_kanji/src/authentication/landing_view.dart";
 import "package:kana_to_kanji/src/core/repositories/user_repository.dart";
-import "package:kana_to_kanji/src/glossary_new/view.dart";
+import "package:kana_to_kanji/src/glossary/view.dart";
 import "package:kana_to_kanji/src/locator.dart";
 import "package:stacked/stacked.dart";
